requirements.txt
termcolor==1.1.0
pydub==0.25.1
ffmpeg-python==0.2.0
moviepy==1.0.3
